Bot code/requirements.txt
aiohappyeyeballs>=2.4.3
aiohttp>=3.11.6
aiosignal>=1.3.1
attrs>=24.2.0
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
frozenlist>=1.5.0
idna>=3.10
multidict>=6.1.0
nextcord>=2.6.0
propcache>=0.2.0
pycparser>=2.22
PyNaCl>=1.5.0
python-dotenv>=1.0.1
requests>=2.32.3
setuptools>=75.6.0
typing_extensions>=4.12.2
urllib3>=2.2.3
yarl>=1.17.2
yt-dlp>=2024.11.18
